replace `MessagePipe` trait with its impl
